+++ build.ps1
@@ -18,13 +18,9 @@ param
 
 $ErrorActionPreference = "Stop"
 
-Write-Host ""
-Write-Host "--------------------------------" -ForegroundColor DarkYellow
-Write-Host " Starting Giraffe build script  " -ForegroundColor DarkYellow
-Write-Host "--------------------------------" -ForegroundColor DarkYellow
-Write-Host ""
+Import-module "$PSScriptRoot/.psscripts/build-functions.ps1" -Force
 
-Import-module "$pwd\.psscripts\build-functions.ps1" -Force
+Write-BuildHeader "Starting Giraffe build script"
 
 if ($ClearOnly.IsPresent)
 {
@@ -32,14 +28,15 @@ if ($ClearOnly.IsPresent)
     return
 }
 
-$giraffe               = ".\src\Giraffe\Giraffe.fsproj"
-$giraffeTests          = ".\tests\Giraffe.Tests\Giraffe.Tests.fsproj"
-$identityApp           = ".\samples\IdentityApp\IdentityApp\IdentityApp.fsproj"
-$jwtApp                = ".\samples\JwtApp\JwtApp\JwtApp.fsproj"
-$sampleApp             = ".\samples\SampleApp\SampleApp\SampleApp.fsproj"
-$sampleAppTests        = ".\samples\SampleApp\SampleApp.Tests\SampleApp.Tests.fsproj"
+$giraffe               = "./src/Giraffe/Giraffe.fsproj"
+$giraffeTests          = "./tests/Giraffe.Tests/Giraffe.Tests.fsproj"
+$identityApp           = "./samples/IdentityApp/IdentityApp/IdentityApp.fsproj"
+$jwtApp                = "./samples/JwtApp/JwtApp/JwtApp.fsproj"
+$sampleApp             = "./samples/SampleApp/SampleApp/SampleApp.fsproj"
+$sampleAppTests        = "./samples/SampleApp/SampleApp.Tests/SampleApp.Tests.fsproj"
 
-Update-AppVeyorBuildVersion $giraffe
+$version = Get-ProjectVersion $giraffe
+Update-AppVeyorBuildVersion $version
 
 if (Test-IsAppVeyorBuildTriggeredByGitTag)
 {
@@ -90,8 +87,4 @@ if ($Pack.IsPresent)
     dotnet-pack $giraffe "-c $configuration"
 }
 
-Write-Host ""
-Write-Host " .~*~*~*~*~*~*~*~*~*~*~*~*~*~*~*~*~*~*~. " -ForegroundColor Green
-Write-Host "  Giraffe build completed successfully!  " -ForegroundColor Green
-Write-Host " '~*~*~*~*~*~*~*~*~*~*~*~*~*~*~*~*~*~*~' " -ForegroundColor Green
-Write-Host ""
+Write-SuccessFooter "Giraffe build completed successfully!"